# Enable SrcAggregator under ugni